Task 0

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,17 +1,16 @@
+#include <stdlib.h>
+#include <string.h>
 #include <stdio.h>
 #include "lists.h"
 
 /**
- * print_list - Prints all elements of a linked list.
- *
- * @h: Pointer to the list
- *
- * Return: The number of nodes.
+ * print_list - check the code for Holberton School students.
+ * @h: name of the list
+ * Return: the number of nodes.
  */
-
 size_t print_list(const list_t *h)
 {
-	size_t  count = 0;
+	int count = 0;
 
 	while (h)
 	{
